+++ src/schedulers.jl
@@ -1,6 +1,7 @@
 module Schedulers
 
 using Base.Threads: nthreads
+using ChunkSplitters: Split, BatchSplit, ScatterSplit
 
 # Used to indicate that a keyword argument has not been set by the user.
 # We don't use Nothing because nothing maybe sometimes be a valid user input (e.g. for init)
@@ -56,10 +57,10 @@ with other multithreaded code.
 - `chunksize::Integer` (default not set)
     * Specifies the desired chunk size (instead of the number of chunks).
     * The options `chunksize` and `nchunks`/`ntasks` are **mutually exclusive** (only one may be a positive integer).
-- `split::Symbol` (default `:batch`):
+- `split::Split` (default `BatchSplit()`):
     * Determines how the collection is divided into chunks (if chunking=true). By default, each chunk consists of contiguous elements and order is maintained.
     * See [ChunkSplitters.jl](https://github.com/JuliaFolds2/ChunkSplitters.jl) for more details and available options.
-    * Beware that for `split=:scatter` the order of elements isn't maintained and a reducer function must not only be associative but also **commutative**!
+    * Beware that for `split=ScatterSplit()` the order of elements isn't maintained and a reducer function must not only be associative but also **commutative**!
 - `chunking::Bool` (default `true`):
     * Controls whether input elements are grouped into chunks (`true`) or not (`false`).
     * For `chunking=false`, the arguments `nchunks`/`ntasks`, `chunksize`, and `split` are ignored and input elements are regarded as "chunks" as is. Hence, there will be one parallel task spawned per input element. Note that, depending on the input, this **might spawn many(!) tasks** and can be costly!
@@ -71,10 +72,10 @@ struct DynamicScheduler{C <: ChunkingMode} <: Scheduler
     threadpool::Symbol
     nchunks::Int
     chunksize::Int
-    split::Symbol
+    split::Split
 
     function DynamicScheduler(threadpool::Symbol, nchunks::Integer, chunksize::Integer,
-            split::Symbol; chunking::Bool = true)
+            split::Split; chunking::Bool = true)
         if !(threadpool in (:default, :interactive))
             throw(ArgumentError("threadpool must be either :default or :interactive"))
         end
@@ -99,7 +100,7 @@ function DynamicScheduler(;
         ntasks::MaybeInteger = NotGiven(), # "alias" for nchunks
         chunksize::MaybeInteger = NotGiven(),
         chunking::Bool = true,
-        split::Symbol = :batch)
+        split::Split = BatchSplit())
     if !chunking
         nchunks = -1
         chunksize = -1
@@ -151,17 +152,17 @@ Isn't well composable with other multithreaded code though.
 - `chunking::Bool` (default `true`):
     * Controls whether input elements are grouped into chunks (`true`) or not (`false`).
     * For `chunking=false`, the arguments `nchunks`/`ntasks`, `chunksize`, and `split` are ignored and input elements are regarded as "chunks" as is. Hence, there will be one parallel task spawned per input element. Note that, depending on the input, this **might spawn many(!) tasks** and can be costly!
-- `split::Symbol` (default `:batch`):
+- `split::Split` (default `BatchSplit()`):
     * Determines how the collection is divided into chunks. By default, each chunk consists of contiguous elements and order is maintained.
     * See [ChunkSplitters.jl](https://github.com/JuliaFolds2/ChunkSplitters.jl) for more details and available options.
-    * Beware that for `split=:scatter` the order of elements isn't maintained and a reducer function must not only be associative but also **commutative**!
+    * Beware that for `split=ScatterSplit()` the order of elements isn't maintained and a reducer function must not only be associative but also **commutative**!
 """
 struct StaticScheduler{C <: ChunkingMode} <: Scheduler
     nchunks::Int
     chunksize::Int
-    split::Symbol
+    split::Split
 
-    function StaticScheduler(nchunks::Integer, chunksize::Integer, split::Symbol;
+    function StaticScheduler(nchunks::Integer, chunksize::Integer, split::Split;
             chunking::Bool = true)
         if !chunking
             C = NoChunking
@@ -183,7 +184,7 @@ function StaticScheduler(;
         ntasks::MaybeInteger = NotGiven(), # "alias" for nchunks
         chunksize::MaybeInteger = NotGiven(),
         chunking::Bool = true,
-        split::Symbol = :batch)
+        split::Split = BatchSplit())
     if !chunking
         nchunks = -1
         chunksize = -1
@@ -238,7 +239,7 @@ some additional overhead.
 - `chunksize::Integer` (default not set)
     * Specifies the desired chunk size (instead of the number of chunks).
     * The options `chunksize` and `nchunks` are **mutually exclusive** (only one may be a positive integer).
-- `split::Symbol` (default `:scatter`):
+- `split::Split` (default `ScatterSplit()`):
     * Determines how the collection is divided into chunks (if chunking=true).
     * See [ChunkSplitters.jl](https://github.com/JuliaFolds2/ChunkSplitters.jl) for more details and available options.
 """
@@ -246,10 +247,10 @@ struct GreedyScheduler{C <: ChunkingMode} <: Scheduler
     ntasks::Int
     nchunks::Int
     chunksize::Int
-    split::Symbol
+    split::Split
 
     function GreedyScheduler(ntasks::Int, nchunks::Integer, chunksize::Integer,
-            split::Symbol; chunking::Bool = false)
+            split::Split; chunking::Bool = false)
         ntasks > 0 || throw(ArgumentError("ntasks must be a positive integer"))
         if !chunking
             C = NoChunking
@@ -271,7 +272,7 @@ function GreedyScheduler(;
         nchunks::MaybeInteger = NotGiven(),
         chunksize::MaybeInteger = NotGiven(),
         chunking::Bool = false,
-        split::Symbol = :scatter)
+        split::Split = ScatterSplit())
     if isgiven(nchunks) || isgiven(chunksize)
         chunking = true
     end